backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
websockets==12.0
pyserial==3.5
numpy==1.26.4
aiosqlite==0.20.0
python-dotenv==1.0.1
